warnings
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@21513 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzflag/HUDRenderer.cxx
@@ -67,7 +67,7 @@ HUDRenderer::HUDRenderer(const BzfDisplay* _display,
 				showCompose(false),
 				showCracks(true),
 				dater(false),
-				lastTimeChange(time(NULL)),
+				lastTimeChange((unsigned long)time(NULL)),
 				triangleCount(0),
 				radarTriangleCount(0)
 {
@@ -736,7 +736,7 @@ void			HUDRenderer::renderStatus(void)
     if (BZDB.eval("timedate") == 2) {
       if (time(NULL) - lastTimeChange >= 2) {
 	dater = !dater;
-	lastTimeChange = time(NULL);
+	lastTimeChange = (unsigned int)time(NULL);
       }
     } else {
       dater = (BZDB.eval("timedate") == 1);
